testui/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
boto3>=1.35
httpx>=0.27
python-dotenv>=1.0
